+++ src/Command.ts
@@ -606,17 +606,16 @@ export default abstract class Command {
         const id = this.getLogItemId(l);
 
         if (title && id) {
-          output.push(`## ${title} (${id})`);
+          output.push(`## ${title} (${id})`, os.EOL, os.EOL);
         }
         else if (title) {
-          output.push(`## ${title}`);
+          output.push(`## ${title}`, os.EOL, os.EOL);
         }
         else if (id) {
-          output.push(`## ${id}`);
+          output.push(`## ${id}`, os.EOL, os.EOL);
         }
 
         output.push(
-          os.EOL, os.EOL,
           `Property | Value`, os.EOL,
           `---------|-------`, os.EOL
         );