+++ src/m365/planner/commands/plan/plan-set.spec.ts
@@ -23,6 +23,7 @@ describe(commands.PLAN_SET, () => {
   const title = 'Plan name';
   const ownerGroupName = 'Group name';
   const ownerGroupId = '00000000-0000-0000-0000-000000000002';
+  const rosterId = 'tYqYlNd6eECmsNhN_fcq85cAGAnd';
   const newTitle = 'New Title';
   const user = 'user@contoso.com';
   const userId = '00000000-0000-0000-0000-000000000000';
@@ -393,6 +394,43 @@ describe(commands.PLAN_SET, () => {
 
     assert(loggerLogSpy.calledWith(outputResponse));
   });
+
+  it('correctly updates planner plan shareWithUserIds with given title and rosterId', async () => {
+    sinon.stub(request, 'get').callsFake(async (opts) => {
+      if (opts.url === `https://graph.microsoft.com/beta/planner/rosters/${rosterId}/plans`) {
+        return Promise.resolve(singlePlansResponse);
+      }
+
+      if (opts.url === `https://graph.microsoft.com/v1.0/planner/plans/${id}`) {
+        return planResponse;
+      }
+
+      if (opts.url === `https://graph.microsoft.com/v1.0/planner/plans/${id}/details`) {
+        return planDetailsResponse;
+      }
+
+      return 'Invalid request';
+    });
+
+    sinon.stub(request, 'patch').callsFake(async (opts) => {
+      if (opts.url === `https://graph.microsoft.com/v1.0/planner/plans/${id}/details`) {
+        return outputResponse;
+      }
+
+      return 'Invalid request';
+    });
+
+    await command.action(logger, {
+      options: {
+        title: title,
+        rosterId: rosterId,
+        shareWithUserIds: shareWithUserIds
+      }
+    });
+
+    assert(loggerLogSpy.calledWith(outputResponse));
+  });
+
 
   it('correctly updates planner plan categories with given id', async () => {
     sinon.stub(request, 'get').callsFake(async (opts) => {