+++ src/m365/purview/commands/retentionevent/retentionevent-list.spec.ts
@@ -1,0 +1,144 @@
+import * as assert from 'assert';
+import * as sinon from 'sinon';
+import { telemetry } from '../../../../telemetry';
+import auth from '../../../../Auth';
+import { Logger } from '../../../../cli/Logger';
+import Command, { CommandError } from '../../../../Command';
+import request from '../../../../request';
+import { pid } from '../../../../utils/pid';
+import { sinonUtil } from '../../../../utils/sinonUtil';
+import commands from '../../commands';
+import { accessToken } from '../../../../utils/accessToken';
+const command: Command = require('./retentionevent-list');
+
+describe(commands.RETENTIONEVENTTYPE_GET, () => {
+
+  //#region Mocked responses
+  const mockResponseArray = [
+    {
+      "displayName": "Retention Event",
+      "description": null,
+      "eventTriggerDateTime": "2023-02-03T13:51:40Z",
+      "eventStatus": null,
+      "lastStatusUpdateDateTime": null,
+      "createdDateTime": "2023-02-03T13:51:40Z",
+      "lastModifiedDateTime": "2023-02-03T13:51:40Z",
+      "id": "7248cfa8-c03a-4ec1-49a4-08db05edc686",
+      "eventQueries": [],
+      "eventPropagationResults": [],
+      "createdBy": {
+        "user": {
+          "id": null,
+          "displayName": "John Doe"
+        }
+      },
+      "lastModifiedBy": {
+        "user": {
+          "id": null,
+          "displayName": "John Doe"
+        }
+      }
+    }
+  ];
+
+  const mockResponse = {
+    "@odata.context": "https://graph.microsoft.com/beta/$metadata#security/triggers/retentionEvents",
+    "@odata.count": 2,
+    "value": mockResponseArray
+  };
+  //#endregion
+
+  let log: string[];
+  let logger: Logger;
+  let loggerLogSpy: sinon.SinonSpy;
+
+  before(() => {
+    sinon.stub(auth, 'restoreAuth').callsFake(() => Promise.resolve());
+    sinon.stub(telemetry, 'trackEvent').callsFake(() => { });
+    sinon.stub(pid, 'getProcessName').callsFake(() => '');
+    auth.service.connected = true;
+    auth.service.accessTokens[auth.defaultResource] = {
+      expiresOn: 'abc',
+      accessToken: 'abc'
+    };
+  });
+
+  beforeEach(() => {
+    log = [];
+    logger = {
+      log: (msg: string) => {
+        log.push(msg);
+      },
+      logRaw: (msg: string) => {
+        log.push(msg);
+      },
+      logToStderr: (msg: string) => {
+        log.push(msg);
+      }
+    };
+    loggerLogSpy = sinon.spy(logger, 'log');
+    (command as any).items = [];
+    sinon.stub(accessToken, 'isAppOnlyAccessToken').callsFake(() => false);
+  });
+
+  afterEach(() => {
+    sinonUtil.restore([
+      accessToken.isAppOnlyAccessToken,
+      request.get
+    ]);
+  });
+
+  after(() => {
+    sinonUtil.restore([
+      auth.restoreAuth,
+      telemetry.trackEvent,
+      pid.getProcessName
+    ]);
+    auth.service.connected = false;
+  });
+
+  it('has correct name', () => {
+    assert.strictEqual(command.name, commands.RETENTIONEVENT_LIST);
+  });
+
+  it('has a description', () => {
+    assert.notStrictEqual(command.description, null);
+  });
+
+  it('defines correct properties for the default output', () => {
+    assert.deepStrictEqual(command.defaultProperties(), ['id', 'displayName', 'eventTriggerDateTime']);
+  });
+
+  it('retrieves retention events', async () => {
+    sinon.stub(request, 'get').callsFake(async (opts) => {
+      if (opts.url === `https://graph.microsoft.com/beta/security/triggers/retentionEvents`) {
+        return mockResponse;
+      }
+
+      throw 'Invalid request';
+    });
+
+    await command.action(logger, { options: { verbose: true } });
+    assert(loggerLogSpy.calledWith(mockResponseArray));
+  });
+
+  it('handles error when retrieving retention events', async () => {
+    sinon.stub(request, 'get').callsFake(async (opts) => {
+      if (opts.url === `https://graph.microsoft.com/beta/security/triggers/retentionEvents`) {
+        throw { error: { error: { message: 'An error has occurred' } } };
+      }
+
+      throw 'Invalid request';
+    });
+
+    await assert.rejects(command.action(logger, { options: {} }), new CommandError('An error has occurred'));
+  });
+
+  it('throws error if something fails using application permissions', async () => {
+    sinonUtil.restore([accessToken.isAppOnlyAccessToken]);
+    sinon.stub(accessToken, 'isAppOnlyAccessToken').callsFake(() => true);
+
+    await assert.rejects(command.action(logger, { options: {} } as any),
+      new CommandError(`This command currently does not support app only permissions.`));
+  });
+});