+++ src/m365/purview/commands/retentioneventtype/retentioneventtype-get.ts
@@ -15,7 +15,7 @@ interface Options extends GlobalOptions {
   id: string;
 }
 
-class PurviewRetentionEventtypeGetCommand extends GraphCommand {
+class PurviewRetentionEventTypeGetCommand extends GraphCommand {
   public get name(): string {
     return commands.RETENTIONEVENTTYPE_GET;
   }
@@ -78,4 +78,4 @@ class PurviewRetentionEventtypeGetCommand extends GraphCommand {
   }
 }
 
-module.exports = new PurviewRetentionEventtypeGetCommand();
+module.exports = new PurviewRetentionEventTypeGetCommand();