+++ src/m365/spo/commands/customaction/customaction-list.spec.ts
@@ -70,64 +70,6 @@ describe(commands.CUSTOMACTION_LIST, () => {
     assert.deepStrictEqual(command.defaultProperties(), ['Name', 'Location', 'Scope', 'Id']);
   });
 
-  it('getCustomActions called once when scope is Web', async () => {
-    const getRequestSpy = sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf('/_api/Web/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    const getCustomActionsSpy = sinon.spy((command as any), 'getCustomActions');
-    const options = {
-      webUrl: 'https://contoso.sharepoint.com',
-      scope: 'Web'
-    };
-
-    try {
-      await command.action(logger, { options: options } as any);
-      assert(getRequestSpy.calledOnce);
-      assert(getCustomActionsSpy.calledWith({
-        webUrl: 'https://contoso.sharepoint.com',
-        scope: 'Web'
-      }));
-      assert(getCustomActionsSpy.calledOnce);
-    }
-    finally {
-      sinonUtil.restore((command as any)['getCustomActions']);
-    }
-  });
-
-  it('getCustomActions called once when scope is Site', async () => {
-    const getRequestSpy = sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf('/_api/Site/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    const getCustomActionsSpy = sinon.spy((command as any), 'getCustomActions');
-    const options = {
-      webUrl: 'https://contoso.sharepoint.com',
-      scope: 'Site'
-    };
-
-    try {
-      await command.action(logger, { options: options } as any);
-      assert(getRequestSpy.calledOnce);
-      assert(getCustomActionsSpy.calledWith({
-        webUrl: 'https://contoso.sharepoint.com',
-        scope: 'Site'
-      }));
-      assert(getCustomActionsSpy.calledOnce);
-    }
-    finally {
-      sinonUtil.restore((command as any)['getCustomActions']);
-    }
-  });
-
   it('returns all properties for output JSON', async () => {
     sinon.stub(request, 'get').callsFake((opts) => {
       if ((opts.url as string).indexOf('/_api/Site/UserCustomActions') > -1) {
@@ -152,86 +94,6 @@ describe(commands.CUSTOMACTION_LIST, () => {
     }
   });
 
-  it('getCustomActions called twice when scope is All', async () => {
-    const getRequestSpy = sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf('/_api/Web/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      if ((opts.url as string).indexOf('/_api/Site/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    const getCustomActionsSpy = sinon.spy((command as any), 'getCustomActions');
-
-    try {
-      await command.action(logger, {
-        options: {
-          debug: true,
-          webUrl: 'https://contoso.sharepoint.com'
-        }
-      });
-      assert(getRequestSpy.calledTwice);
-      assert(getCustomActionsSpy.calledTwice);
-    }
-    finally {
-      sinonUtil.restore((command as any)['getCustomActions']);
-    }
-  });
-
-  it('searchAllScopes called when scope is All', async () => {
-    sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf('/_api/Web/UserCustomActions') > -1) {
-        return Promise.resolve('abc');
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    const searchAllScopesSpy = sinon.spy((command as any), 'searchAllScopes');
-    const options = {
-      webUrl: 'https://contoso.sharepoint.com',
-      scope: "All"
-    };
-
-    try {
-      await assert.rejects(command.action(logger, { options: options } as any));
-      assert(searchAllScopesSpy.calledWith(sinon.match(
-        {
-          webUrl: 'https://contoso.sharepoint.com'
-        })));
-      assert(searchAllScopesSpy.calledOnce);
-    }
-    finally {
-      sinonUtil.restore((command as any)['searchAllScopes']);
-    }
-  });
-
-  it('searchAllScopes correctly handles no custom actions when All scope specified', async () => {
-    sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf('/_api/Web/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      if ((opts.url as string).indexOf('/_api/Site/UserCustomActions') > -1) {
-        return Promise.resolve({ value: [] });
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    await command.action(logger, {
-      options: {
-        verbose: false,
-        webUrl: 'https://contoso.sharepoint.com',
-        scope: 'All'
-      }
-    });
-    assert(loggerLogSpy.notCalled);
-  });
 
   it('correctly handles no custom actions when All scope specified (verbose)', async () => {
     sinon.stub(request, 'get').callsFake((opts) => {
@@ -261,6 +123,9 @@ describe(commands.CUSTOMACTION_LIST, () => {
     sinon.stub(request, 'get').callsFake((opts) => {
       if ((opts.url as string).indexOf('/_api/Web/UserCustomActions') > -1) {
         return Promise.reject(err);
+      }
+      else if ((opts.url as string).indexOf('/_api/Site/UserCustomActions') > -1) {
+        return Promise.resolve({ value: [] });
       }
 
       return Promise.reject('Invalid request');